LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/rendering/widgets/button-layout/abspos-expected.txt

testtesttest

FAIL abspos button with auto width, non-auto left/right (rtl) assert_equals: offsetLeft expected 661 but got 102
FAIL abspos button with auto width, non-auto left/right (ltr) assert_equals: clientWidth expected 33 but got 592
